requirements.txt
asgiref==3.5.2
autopep8==1.7.0
Django==4.0.5
django-environ==0.9.0
django-polymorphic==3.1.0
djangorestframework==3.13.1
django-rest-swagger==2.2.0
pycodestyle==2.9.1
pytz==2022.2.1
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
mysqlclient==2.1.1
django-filter==22.1
django-cors-headers==3.13.0
psycopg2==2.9.4
djangorestframework-serializer-extensions==2.0.1